requirements.txt
# Core dependencies
anthropic>=0.40.0
openai>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
pyyaml>=6.0

# Benchmarking
tenacity>=8.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0

# Visualization
matplotlib>=3.8.0
plotly>=5.18.0
kaleido>=0.2.0

# Utilities
python-dotenv>=1.0.0
colorlog>=6.8.0
